pyproject.toml
[tool.poetry]
name = "eazy"
version = "0.1.0"
description = ""
authors = ["akshitkrnagpal <akshitkrnagpal@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7"
requests = "^2.28.2"
python-dotenv = "^0.21.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"
pytest-mock = "*"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
